requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
httpx
pinecone-client
